Refactors parse-ast
