+++ lib/parse/javascript/parse-ast.js
@@ -1,92 +1,39 @@
-const { get } = require('kompis');
+const { assign, get, id, isSome, map, match } = require('kompis');
+const { otherwise, Pipe, reduce } = require('kompis');
+
 const traverse = require('@babel/traverse').default;
 const t = require('@babel/types');
 
-const matchNode = require('../../utils/match-node');
-const parseObjectMethod = require('./parse-object-method');
+const { throwError } = require('../../utils/error-handling');
+const parseCallExpression = require('./parse-call-expression');
 
-const parseArrayElement = node => {
-  return matchNode(
-    node,
-    {
-      StringLiteral: node => node.value,
-      NumericLiteral: node => node.value
-    },
-    () => {
-      throw new Error(`Unsupported ${node.type} in PropTypes.oneOf`);
-    }
-  );
-};
+const parseArrayElement = match(
+  [isSome(t.isStringLiteral, t.isNumericLiteral), get('value')],
+  [otherwise, node => throwError(`Unsupported ${node.type} in PropTypes.oneOf`)]
+);
 
-const parseType = (node, meta = {}) => {
-  return matchNode(
-    node,
-    {
-      ArrayExpression: ({ elements }) => elements.map(parseArrayElement),
-      // Arguments to PropTypes methods only accept one argument (destructured into 'argument')
-      CallExpression: path => {
-        const {
-          callee,
-          arguments: [argument]
-        } = path;
+const parseProperty = meta => p => ({
+  [p.key.name]: parseType(p.value, get(`argument.${p.key.name}`)(meta))
+});
 
-        // Handle component references in 'shape'
-        if (
-          ['shape', 'exact'].includes(callee.name) &&
-          t.isMemberExpression(argument) &&
-          t.isIdentifier(argument.property, { name: 'propTypes' })
-        ) {
-          return parseType(argument.object, meta);
-        }
+const parseMemberExpression = meta => node => ({
+  ...parseType(node.object, meta),
+  isRequired: t.isIdentifier(node.property, { name: 'isRequired' })
+});
 
-        if (meta.type === 'exclude') {
-          return;
-        }
-
-        // 'instanceOf' is included here because it is automatically stripped in 'transform-prop-types.js' and should not throw an error
-        if (
-          ['arrayOf', 'exact', 'instanceOf', 'oneOf', 'shape'].includes(
-            callee.name
-          )
-        ) {
-          return {
-            type: callee.name,
-            argument: parseType(argument, meta)
-          };
-        }
-
-        // Object.keys, Object.values
-        if (
-          t.isMemberExpression(callee) &&
-          t.isIdentifier(callee.object, { name: 'Object' })
-        ) {
-          return parseObjectMethod(path);
-        }
-
-        throw new Error(`Invalid function call '${callee.name}'`);
-      },
-      Identifier: ({ name }) => ({ type: name }),
-      ObjectExpression: ({ properties }) =>
-        properties.reduce(
-          (accum, property) => ({
-            ...accum,
-            [property.key.name]: parseType(
-              property.value,
-              get(`argument.${property.key.name}`)(meta)
-            )
-          }),
-          {}
-        ),
-      MemberExpression: node => ({
-        ...parseType(node.object, meta),
-        isRequired: t.isIdentifier(node.property, { name: 'isRequired' })
-      })
-    },
-    () => {
-      return node;
-    }
-  );
-};
+function parseType(node, meta = {}) {
+  return match(
+    [t.isArrayExpression, Pipe(get('elements'), map(parseArrayElement))],
+    [t.isCallExpression, parseCallExpression(meta, parseType)],
+    [t.isIdentifier, ({ name }) => ({ type: name })],
+    [
+      t.isObjectExpression,
+      Pipe(get('properties', []), reduce(assign, {}, parseProperty(meta)))
+    ],
+    [t.isMemberExpression, parseMemberExpression(meta)],
+    [otherwise, id]
+  )(node);
+}
 
 module.exports = function(syntaxTree, propTypesMeta = {}) {
   let propTypes;
